카드 합체 놀이(15903) 성공

+++ src/baekjoon/p15903/Main2.java
@@ -8,7 +8,6 @@ import java.util.StringTokenizer;
 
 public class Main2 {
 	// 매 합체마다 카드의 총 합은 (x + y)만큼 늘어남
-	
 	public static void main(String[] args) throws IOException {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		StringTokenizer st = new StringTokenizer(br.readLine(), " ");
@@ -16,16 +15,16 @@ public class Main2 {
 		int m = Integer.parseInt(st.nextToken());
 		st = new StringTokenizer(br.readLine(), " ");
 		// 가장 작은 원소들 부터 뽑을 수 있도록 우선순위 큐 사용
-		PriorityQueue<Integer> pq = new PriorityQueue<Integer>();
+		PriorityQueue<Long> pq = new PriorityQueue<Long>();
 		// 제일 처음 카드들의 합을 구해놓는다.
-		int sum = 0;
+		long sum = 0;
 		for (int i = 0; i < n; i++) {
 			int a = Integer.parseInt(st.nextToken());
 			sum += a;
-			pq.offer(a);
+			pq.offer((long)a);
 		}
 		for (int i = 0; i < m; i++) {
-			int add = (pq.poll() + pq.poll());
+			long add = (pq.poll() + pq.poll());
 			pq.offer(add); pq.offer(add);
 			sum += add;
 		}
